Move memory management classes (KernelArena, KernelScope, KernelScopedObject) to a separate file.
